export notifs parsing
