+++ gnmidiff/gnmiparse/testdata/getresponse.textproto
@@ -1,0 +1,291 @@
+notification: {
+  timestamp: 1676419100456944135
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "ssh-server"
+      }
+      elem: {
+        name: "state"
+      }
+      elem: {
+        name: "protocol-version"
+      }
+    }
+    val: {
+      string_val: "V2"
+    }
+  }
+}
+notification: {
+  timestamp: 1676420328291197426
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "config"
+      }
+      elem: {
+        name: "hostname"
+      }
+    }
+    val: {
+      string_val: "rosesarered"
+    }
+  }
+}
+notification: {
+  timestamp: 1676419100456944135
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "ntp"
+      }
+      elem: {
+        name: "state"
+      }
+      elem: {
+        name: "enabled"
+      }
+    }
+    val: {
+      bool_val: false
+    }
+  }
+}
+notification: {
+  timestamp: 1676419100456944135
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "ntp"
+      }
+      elem: {
+        name: "state"
+      }
+      elem: {
+        name: "enable-ntp-auth"
+      }
+    }
+    val: {
+      bool_val: false
+    }
+  }
+}
+notification: {
+  timestamp: 1676419100456944135
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "ssh-server"
+      }
+      elem: {
+        name: "state"
+      }
+      elem: {
+        name: "enable"
+      }
+    }
+    val: {
+      bool_val: true
+    }
+  }
+}
+notification: {
+  timestamp: 1676420328448197153
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "state"
+      }
+      elem: {
+        name: "hostname"
+      }
+    }
+    val: {
+      string_val: "rosesarered"
+    }
+  }
+}
+notification: {
+  timestamp: 1676419100459254468
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "state"
+      }
+      elem: {
+        name: "boot-time"
+      }
+    }
+    val: {
+      uint_val: 1676419100459308639
+    }
+  }
+}
+notification: {
+  timestamp: 1676419100456944135
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "telnet-server"
+      }
+      elem: {
+        name: "state"
+      }
+      elem: {
+        name: "enable"
+      }
+    }
+    val: {
+      bool_val: false
+    }
+  }
+}
+notification: {
+  timestamp: 1676422427135895887
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "config"
+      }
+      elem: {
+        name: "hostname"
+      }
+    }
+    val: {
+      string_val: "rosesareredd"
+    }
+  }
+}
+notification: {
+  timestamp: 1676422427269965151
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "state"
+      }
+      elem: {
+        name: "hostname"
+      }
+    }
+    val: {
+      string_val: "rosesareredd"
+    }
+  }
+}
+notification: {
+  timestamp: 1676422434342310772
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "config"
+      }
+      elem: {
+        name: "hostname"
+      }
+    }
+    val: {
+      string_val: "rosesarered"
+    }
+  }
+}
+notification: {
+  timestamp: 1676422434479082363
+  prefix: {
+    origin: "openconfig"
+    target: "fakedut"
+  }
+  update: {
+    path: {
+      elem: {
+        name: "system"
+      }
+      elem: {
+        name: "state"
+      }
+      elem: {
+        name: "hostname"
+      }
+    }
+    val: {
+      string_val: "rosesarered"
+    }
+  }
+}
